Move `DownloadManager` to a separate file, display alert on download completion and fix minor bug
